+++ src/components/Nav/index.tsx
@@ -1,0 +1,33 @@
+import React, { useContext } from 'react'
+import { Nav, ConnectorId } from '@pancakeswap-libs/uikit'
+import { useWeb3React } from '@web3-react/core'
+import { allLanguages } from 'constants/localisation/languageCodes'
+import { LanguageContext } from 'hooks/LanguageContext'
+import useTheme from 'hooks/useTheme'
+import { injected, walletconnect } from 'connectors'
+
+const Menu = () => {
+  const { account, activate, deactivate } = useWeb3React()
+  const { selectedLanguage, setSelectedLanguage } = useContext(LanguageContext)
+  const { isDark, toggleTheme } = useTheme()
+
+  return (
+    <Nav
+      account={account as string}
+      login={(connectorId: ConnectorId) => {
+        if (connectorId === 'walletconnect') {
+          return activate(walletconnect)
+        }
+        return activate(injected)
+      }}
+      logout={deactivate}
+      isDark={isDark}
+      toggleTheme={toggleTheme}
+      currentLang={selectedLanguage?.code || ''}
+      langs={allLanguages}
+      setLang={setSelectedLanguage}
+    />
+  )
+}
+
+export default Menu
